Return an empty Hash instead of array for empty arrays
Since PHPs serialized associative arrays are the same as arrays but with
a string instead of an integer for the index. It isn't possible to tell
the difference between an empty array and an empty associative array.
To make empty flash messages work, we will return an empty hash instead
of an array.
https://github.com/rails/rails/blob/v5.1.5/actionpack/lib/action_dispatch/middleware/flash.rb#L125